Switch statistic.rb into an override
